refactor: extract model settings into dedicated component
Split the model tab UI out of Settings.tsx (5162 → 3502 lines) into:
- settings/ModelSettingsSection.tsx: presentational component for model tab
- settings/modelProviderUtils.ts: shared provider types and utility functions
